requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langgraph==0.2.34
langchain==0.3.7
langchain-openai==0.2.5
langchain-groq==1.1.2
langchain-community==0.3.5
langchain-chroma==0.1.4
langchain-text-splitters==0.3.2
chromadb==0.5.20
pymupdf==1.24.13
openai==1.54.3
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
